[PICA] Switch to xxh3 by default
